+++ src/tx-builder/services/Faucet.ts
@@ -1,8 +1,4 @@
-import {
-  commonContractAddressBetweenMarketsV2,
-  DEFAULT_NULL_VALUE_ON_TX,
-  enabledNetworksByService,
-} from '../config';
+import { DEFAULT_NULL_VALUE_ON_TX } from '../config';
 import {
   IFaucet,
   IMinter,
@@ -14,9 +10,8 @@ import {
   Configuration,
   eEthereumTxType,
   EthereumTransactionTypeExtended,
-  tEthereumAddress,
+  LendingPoolMarketConfig,
   transactionType,
-  tStringDecimalUnits,
 } from '../types';
 import { FaucetParamsType } from '../types/FaucetMethodTypes';
 import { mintAmountsPerToken } from '../utils/parsings';
@@ -31,15 +26,21 @@ export default class FaucetService
 
   readonly faucetContract: IFaucet;
 
-  constructor(config: Configuration) {
+  readonly faucetConfig: LendingPoolMarketConfig | undefined;
+
+  constructor(
+    config: Configuration,
+    faucetConfig: LendingPoolMarketConfig | undefined
+  ) {
     super(config, IMinter__factory);
 
-    const { provider, network } = this.config;
+    this.faucetConfig = faucetConfig;
 
-    const { FAUCET } = commonContractAddressBetweenMarketsV2[network];
-    this.faucetAddress = FAUCET;
+    const { provider } = this.config;
 
-    if (enabledNetworksByService.faucet.indexOf(network) > -1) {
+    this.faucetAddress = this.faucetConfig?.FAUCET || '';
+
+    if (this.faucetAddress !== '') {
       this.faucetContract = IFaucet__factory.connect(
         this.faucetAddress,
         provider
@@ -54,13 +55,12 @@ export default class FaucetService
     { userAddress, reserve, tokenSymbol }: FaucetParamsType
   ): Promise<EthereumTransactionTypeExtended[]> {
     const amount: string = mintAmountsPerToken[tokenSymbol];
-    const txValue = await this.getTxValue(reserve, amount);
 
     const txCallback: () => Promise<transactionType> = this.generateTxCallback({
       rawTxMethod: () =>
         this.faucetContract.populateTransaction.mint(reserve, amount),
       from: userAddress,
-      value: txValue,
+      value: DEFAULT_NULL_VALUE_ON_TX,
     });
 
     return [
@@ -70,17 +70,5 @@ export default class FaucetService
         gas: this.generateTxPriceEstimation([], txCallback),
       },
     ];
-  }
-
-  private async getTxValue(
-    token: tEthereumAddress,
-    amount: tStringDecimalUnits
-  ): Promise<tStringDecimalUnits> {
-    const minterAddress: string = await this.faucetContract.getMinter(token);
-    const minterContract: IMinter = this.getContractInstance(minterAddress);
-
-    const isEthRequired: boolean = await minterContract.isEthRequired();
-
-    return isEthRequired ? amount : DEFAULT_NULL_VALUE_ON_TX;
   }
 }