+++ src/tx-builder/services/RepayWithCollateralAdapter.ts
@@ -1,4 +1,3 @@
-import { commonContractAddressBetweenMarketsV2 } from '../config';
 import {
   IRepayWithCollateral,
   IRepayWithCollateral__factory,
@@ -8,6 +7,8 @@ import {
   Configuration,
   eEthereumTxType,
   EthereumTransactionTypeExtended,
+  LendingPoolMarketConfig,
+  ProtocolAction,
   transactionType,
 } from '../types';
 import { RepayWithCollateralType } from '../types/RepayWithCollateralMethodTypes';
@@ -20,14 +21,17 @@ export default class RepayWithCollateralAdapterService
   implements RepayWithCollateralAdapterInterface {
   readonly repayWithCollateralAddress: string;
 
-  constructor(config: Configuration) {
+  readonly repayWithCollateralConfig: LendingPoolMarketConfig | undefined;
+
+  constructor(
+    config: Configuration,
+    repayWithCollateralConfig: LendingPoolMarketConfig | undefined
+  ) {
     super(config, IRepayWithCollateral__factory);
+    this.repayWithCollateralConfig = repayWithCollateralConfig;
 
-    const {
-      REPAY_WITH_COLLATERAL_ADAPTER,
-    } = commonContractAddressBetweenMarketsV2[this.config.network];
-
-    this.repayWithCollateralAddress = REPAY_WITH_COLLATERAL_ADAPTER;
+    this.repayWithCollateralAddress =
+      this.repayWithCollateralConfig?.REPAY_WITH_COLLATERAL_ADAPTER || '';
   }
 
   @RepayWithCollateralValidator
@@ -46,7 +50,8 @@ export default class RepayWithCollateralAdapterService
       debtRateMode,
       permit,
       useEthPath,
-    }: RepayWithCollateralType
+    }: RepayWithCollateralType,
+    txs?: EthereumTransactionTypeExtended[]
   ): EthereumTransactionTypeExtended {
     const repayWithCollateralContract: IRepayWithCollateral = this.getContractInstance(
       this.repayWithCollateralAddress
@@ -69,7 +74,11 @@ export default class RepayWithCollateralAdapterService
     return {
       tx: txCallback,
       txType: eEthereumTxType.DLP_ACTION,
-      gas: this.generateTxPriceEstimation([], txCallback),
+      gas: this.generateTxPriceEstimation(
+        txs || [],
+        txCallback,
+        ProtocolAction.repayCollateral
+      ),
     };
   }
 }