+++ src/tx-builder/services/v2/LendingPool.ts
@@ -1,10 +1,8 @@
-import { BigNumber, constants, utils, BigNumberish, BytesLike } from 'ethers';
+import { constants, utils, BigNumberish, BytesLike } from 'ethers';
 import {
   API_ETH_MOCK_ADDRESS,
-  commonContractAddressBetweenMarketsV2,
   DEFAULT_APPROVE_AMOUNT,
   MAX_UINT_AMOUNT,
-  distinctContractAddressBetweenMarketsV2,
   SURPLUS,
 } from '../../config';
 import { ILendingPool, ILendingPool__factory } from '../../contract-types';
@@ -16,15 +14,20 @@ import {
   eEthereumTxType,
   EthereumTransactionTypeExtended,
   InterestRate,
-  Market,
   ProtocolAction,
   TokenMetadataType,
   transactionType,
   tStringDecimalUnits,
   tEthereumAddress,
+  LendingPoolMarketConfig,
 } from '../../types';
 import { getTxValue, parseNumber } from '../../utils/parsings';
-import { LPValidator } from '../../validators/methodValidators';
+import {
+  LPFlashLiquidationValidator,
+  LPRepayWithCollateralValidator,
+  LPSwapCollateralValidator,
+  LPValidator,
+} from '../../validators/methodValidators';
 import {
   LPBorrowParamsType,
   LPDepositParamsType,
@@ -85,7 +88,7 @@ const buildParaSwapLiquiditySwapParams = (
 export default class LendingPool
   extends BaseService<ILendingPool>
   implements LendingPoolInterface {
-  readonly market: Market;
+  readonly market: string;
 
   readonly erc20Service: IERC20ServiceInterface;
 
@@ -101,6 +104,14 @@ export default class LendingPool
 
   readonly baseDebtTokenService: BaseDebtTokenInterface;
 
+  readonly lendingPoolConfig: LendingPoolMarketConfig | undefined;
+
+  readonly flashLiquidationAddress: string;
+
+  readonly swapCollateralAddress: string;
+
+  readonly repayWithCollateralAddress: string;
+
   constructor(
     config: Configuration,
     erc20Service: IERC20ServiceInterface,
@@ -108,8 +119,13 @@ export default class LendingPool
     wethGatewayService: WETHGatewayInterface,
     liquiditySwapAdapterService: LiquiditySwapAdapterInterface,
     repayWithCollateralAdapterService: RepayWithCollateralAdapterInterface,
+<<<<<<< HEAD
     baseDebtTokenService: BaseDebtTokenInterface,
     market: Market
+=======
+    market: string,
+    lendingPoolConfig: LendingPoolMarketConfig | undefined
+>>>>>>> master
   ) {
     super(config, ILendingPool__factory);
     this.erc20Service = erc20Service;
@@ -119,12 +135,19 @@ export default class LendingPool
     this.repayWithCollateralAdapterService = repayWithCollateralAdapterService;
     this.baseDebtTokenService = baseDebtTokenService;
     this.market = market;
+    this.lendingPoolConfig = lendingPoolConfig;
 
-    const { network } = this.config;
-    this.lendingPoolAddress =
-      distinctContractAddressBetweenMarketsV2[this.market][
-        network
-      ].LENDINGPOOL_ADDRESS;
+    const {
+      LENDING_POOL,
+      FLASH_LIQUIDATION_ADAPTER,
+      REPAY_WITH_COLLATERAL_ADAPTER,
+      SWAP_COLLATERAL_ADAPTER,
+    } = this.lendingPoolConfig || {};
+
+    this.lendingPoolAddress = LENDING_POOL || '';
+    this.flashLiquidationAddress = FLASH_LIQUIDATION_ADAPTER || '';
+    this.swapCollateralAddress = SWAP_COLLATERAL_ADAPTER || '';
+    this.repayWithCollateralAddress = REPAY_WITH_COLLATERAL_ADAPTER || '';
   }
 
   @LPValidator
@@ -564,7 +587,7 @@ export default class LendingPool
     return txs;
   }
 
-  @LPValidator
+  @LPSwapCollateralValidator
   public async swapCollateral(
     @IsEthAddress('user')
     @IsEthAddress('fromAsset')
@@ -600,14 +623,10 @@ export default class LendingPool
       s: '0x0000000000000000000000000000000000000000000000000000000000000000',
     };
 
-    const { SWAP_COLLATERAL_ADAPTER } = commonContractAddressBetweenMarketsV2[
-      this.config.network
-    ];
-
     const approved: boolean = await this.erc20Service.isApproved(
       fromAToken,
       user,
-      SWAP_COLLATERAL_ADAPTER,
+      this.swapCollateralAddress,
       fromAmount
     );
 
@@ -615,7 +634,7 @@ export default class LendingPool
       const approveTx: EthereumTransactionTypeExtended = this.erc20Service.approve(
         user,
         fromAToken,
-        SWAP_COLLATERAL_ADAPTER,
+        this.swapCollateralAddress,
         constants.MaxUint256.toString()
       );
 
@@ -653,15 +672,6 @@ export default class LendingPool
     );
 
     if (flash) {
-      const FLASHLOAN_PREMIUM_TOTAL: BigNumber = await lendingPoolContract.FLASHLOAN_PREMIUM_TOTAL();
-      const convertedAmountNoFees: string = BigNumber.from(convertedAmount)
-        .sub(
-          BigNumber.from(convertedAmount)
-            .mul(FLASHLOAN_PREMIUM_TOTAL)
-            .div(10000)
-        )
-        .toString();
-
       const amountWithSurplus: string = (
         Number(fromAmount) +
         (Number(fromAmount) * Number(SURPLUS)) / 100
@@ -676,9 +686,9 @@ export default class LendingPool
         {
           rawTxMethod: () =>
             lendingPoolContract.populateTransaction.flashLoan(
-              SWAP_COLLATERAL_ADAPTER,
+              this.swapCollateralAddress,
               [fromAsset],
-              swapAll ? [convertedAmountWithSurplus] : [convertedAmountNoFees],
+              swapAll ? [convertedAmountWithSurplus] : [convertedAmount],
               [0], // interest rate mode to NONE for flashloan to not open debt
               onBehalfOf || user,
               params,
@@ -712,14 +722,15 @@ export default class LendingPool
         swapCallData,
         augustus,
         permitParams,
-      }
+      },
+      txs
     );
 
     txs.push(swapAndDepositTx);
     return txs;
   }
 
-  @LPValidator
+  @LPRepayWithCollateralValidator
   public async repayWithCollateral(
     @IsEthAddress('user')
     @IsEthAddress('fromAsset')
@@ -754,14 +765,10 @@ export default class LendingPool
       s: '0x0000000000000000000000000000000000000000000000000000000000000000',
     };
 
-    const {
-      REPAY_WITH_COLLATERAL_ADAPTER,
-    } = commonContractAddressBetweenMarketsV2[this.config.network];
-
     const approved: boolean = await this.erc20Service.isApproved(
       fromAToken,
       user,
-      REPAY_WITH_COLLATERAL_ADAPTER,
+      this.repayWithCollateralAddress,
       repayWithAmount
     );
 
@@ -769,7 +776,7 @@ export default class LendingPool
       const approveTx: EthereumTransactionTypeExtended = this.erc20Service.approve(
         user,
         fromAToken,
-        REPAY_WITH_COLLATERAL_ADAPTER,
+        this.repayWithCollateralAddress,
         constants.MaxUint256.toString()
       );
 
@@ -831,7 +838,7 @@ export default class LendingPool
         {
           rawTxMethod: () =>
             lendingPoolContract.populateTransaction.flashLoan(
-              REPAY_WITH_COLLATERAL_ADAPTER,
+              this.repayWithCollateralAddress,
               [assetToRepay],
               [convertedRepayAmount],
               [0], // interest rate mode to NONE for flashloan to not open debt
@@ -866,7 +873,8 @@ export default class LendingPool
         debtRateMode: numericInterestRate,
         permit: permitParams,
         useEthPath,
-      }
+      },
+      txs
     );
 
     txs.push(swapAndRepayTx);
@@ -874,7 +882,7 @@ export default class LendingPool
     return txs;
   }
 
-  @LPValidator
+  @LPFlashLiquidationValidator
   public async flashLiquidation(
     @IsEthAddress('user')
     @IsEthAddress('collateralAsset')
@@ -899,10 +907,6 @@ export default class LendingPool
     };
 
     const txs: EthereumTransactionTypeExtended[] = [];
-
-    const { FLASHLIQUIDATION } = commonContractAddressBetweenMarketsV2[
-      this.config.network
-    ];
 
     const lendingPoolContract: ILendingPool = this.getContractInstance(
       this.lendingPoolAddress
@@ -936,7 +940,7 @@ export default class LendingPool
     const txCallback: () => Promise<transactionType> = this.generateTxCallback({
       rawTxMethod: () =>
         lendingPoolContract.populateTransaction.flashLoan(
-          FLASHLIQUIDATION,
+          this.flashLiquidationAddress,
           [borrowedAsset],
           [flashBorrowAmount],
           [0],