+++ src/tx-builder/v2.ts
@@ -1,5 +1,5 @@
 import { providers } from 'ethers';
-import { Network, Market, DefaultProviderKeys } from './types';
+import { Network, DefaultProviderKeys, TxBuilderConfig } from './types';
 import TxBuilderInterface from './interfaces/TxBuilder';
 import LendingPoolInterface from './interfaces/v2/LendingPool';
 import LendingPool from './services/v2/LendingPool';
@@ -19,69 +19,114 @@ import GovernanceDelegationTokenService from './services/v2/GovernanceDelegation
 
 export default class TxBuilder
   extends BaseTxBuilder
-  implements TxBuilderInterface {
+  implements TxBuilderInterface
+{
   readonly lendingPools: {
     [market: string]: LendingPoolInterface;
   };
 
+  readonly wethGateways: {
+    [market: string]: WETHGatewayInterface;
+  };
+
+  readonly swapCollateralAdapters: {
+    [market: string]: LiquiditySwapAdapterInterface;
+  };
+
+  readonly repayWithCollateralAdapters: {
+    [market: string]: RepayWithCollateralAdapterInterface;
+  };
+
   readonly baseDebtTokenService: BaseDebtTokenInterface;
-
-  readonly liquiditySwapAdapterService: LiquiditySwapAdapterInterface;
-
-  readonly repayWithCollateralAdapterService: RepayWithCollateralAdapterInterface;
 
   public aaveGovernanceV2Service: AaveGovernanceV2Interface;
 
   public governanceDelegationTokenService: GovernanceDelegationTokenInterface;
 
-  public wethGatewayService: WETHGatewayInterface;
-
   constructor(
     network: Network = Network.mainnet,
-    injectedProvider?:
-      | providers.ExternalProvider
-      | providers.Web3Provider
-      | string
-      | undefined,
-    defaultProviderKeys?: DefaultProviderKeys
+    injectedProvider?: providers.Provider | string | undefined,
+    defaultProviderKeys?: DefaultProviderKeys,
+    config?: TxBuilderConfig
   ) {
-    super(network, injectedProvider, defaultProviderKeys);
+    super(network, injectedProvider, defaultProviderKeys, config);
 
+    this.wethGateways = {};
     this.lendingPools = {};
+    this.swapCollateralAdapters = {};
+    this.repayWithCollateralAdapters = {};
     this.baseDebtTokenService = new BaseDebtTokenService(
       this.configuration,
       this.erc20Service
     );
-    this.wethGatewayService = new WETHGatewayService(
-      this.configuration,
-      this.baseDebtTokenService,
-      this.erc20Service
-    );
-    this.liquiditySwapAdapterService = new LiquiditySwapAdapterService(
-      this.configuration
-    );
-    this.repayWithCollateralAdapterService = new RepayWithCollateralAdapterService(
-      this.configuration
-    );
+
     this.aaveGovernanceV2Service = new AaveGovernanceV2Service(
-      this.configuration
+      this.configuration,
+      this.txBuilderConfig.governance?.[network]
     );
-    this.governanceDelegationTokenService = new GovernanceDelegationTokenService(
-      this.configuration
-    );
+
+    this.governanceDelegationTokenService =
+      new GovernanceDelegationTokenService(this.configuration);
   }
 
-  public getLendingPool = (market: Market): LendingPoolInterface => {
+  public getRepayWithCollateralAdapter = (
+    market: string
+  ): RepayWithCollateralAdapterInterface => {
+    const { network } = this.configuration;
+
+    if (!this.repayWithCollateralAdapters[market]) {
+      this.repayWithCollateralAdapters[market] =
+        new RepayWithCollateralAdapterService(
+          this.configuration,
+          this.txBuilderConfig.lendingPool?.[network]?.[market]
+        );
+    }
+
+    return this.repayWithCollateralAdapters[market];
+  };
+
+  public getSwapCollateralAdapter = (
+    market: string
+  ): LiquiditySwapAdapterInterface => {
+    const { network } = this.configuration;
+
+    if (!this.swapCollateralAdapters[market]) {
+      this.swapCollateralAdapters[market] = new LiquiditySwapAdapterService(
+        this.configuration,
+        this.txBuilderConfig.lendingPool?.[network]?.[market]
+      );
+    }
+
+    return this.swapCollateralAdapters[market];
+  };
+
+  public getWethGateway = (market: string): WETHGatewayInterface => {
+    const { network } = this.configuration;
+    if (!this.wethGateways[market]) {
+      this.wethGateways[market] = new WETHGatewayService(
+        this.configuration,
+        this.baseDebtTokenService,
+        this.erc20Service,
+        this.txBuilderConfig.lendingPool?.[network]?.[market]
+      );
+    }
+
+    return this.wethGateways[market];
+  };
+
+  public getLendingPool = (market: string): LendingPoolInterface => {
+    const { network } = this.configuration;
     if (!this.lendingPools[market]) {
       this.lendingPools[market] = new LendingPool(
         this.configuration,
         this.erc20Service,
         this.synthetixService,
-        this.wethGatewayService,
-        this.liquiditySwapAdapterService,
-        this.repayWithCollateralAdapterService,
+        this.getWethGateway(market),
+        this.getSwapCollateralAdapter(market),
+        this.getRepayWithCollateralAdapter(market),
         this.baseDebtTokenService,
-        market
+        market,
+        this.txBuilderConfig.lendingPool?.[network]?.[market]
       );
     }