+++ src/tx-builder/validators/methodValidators.ts
@@ -1,16 +1,113 @@
 /* eslint-disable prefer-rest-params */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { Network } from '../types';
 import {
   amount0OrPositiveValidator,
   amountGtThan0OrMinus1,
   amountGtThan0Validator,
-  // isEthAddressArrayValidator,
   isEthAddressOrEnsValidator,
   isEthAddressValidator,
   optionalValidator,
 } from './validations';
-import { enabledNetworksByService } from '../config';
+import { utils } from 'ethers';
+
+export function LPFlashLiquidationValidator(
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  target: any,
+  propertyName: string,
+  descriptor: TypedPropertyDescriptor<any>
+): any {
+  const method = descriptor.value;
+  // eslint-disable-next-line no-param-reassign
+  descriptor.value = function () {
+    const { LENDING_POOL, FLASH_LIQUIDATION_ADAPTER } =
+      this.lendingPoolConfig || {};
+
+    if (
+      !utils.isAddress(LENDING_POOL) ||
+      !FLASH_LIQUIDATION_ADAPTER ||
+      !utils.isAddress(FLASH_LIQUIDATION_ADAPTER)
+    ) {
+      console.error(
+        `[LPFlahsLiquidationValidator] You need to pass valid addresses`
+      );
+      return [];
+    }
+
+    isEthAddressValidator(target, propertyName, arguments);
+
+    amountGtThan0Validator(target, propertyName, arguments);
+
+    amountGtThan0OrMinus1(target, propertyName, arguments);
+
+    return method?.apply(this, arguments);
+  };
+}
+
+export function LPRepayWithCollateralValidator(
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  target: any,
+  propertyName: string,
+  descriptor: TypedPropertyDescriptor<any>
+): any {
+  const method = descriptor.value;
+  // eslint-disable-next-line no-param-reassign
+  descriptor.value = function () {
+    const { LENDING_POOL, REPAY_WITH_COLLATERAL_ADAPTER } =
+      this.lendingPoolConfig || {};
+
+    if (
+      !utils.isAddress(LENDING_POOL) ||
+      !REPAY_WITH_COLLATERAL_ADAPTER ||
+      !utils.isAddress(REPAY_WITH_COLLATERAL_ADAPTER)
+    ) {
+      console.error(
+        `[LPRepayWithCollateralValidator] You need to pass valid addresses`
+      );
+      return [];
+    }
+
+    isEthAddressValidator(target, propertyName, arguments);
+
+    amountGtThan0Validator(target, propertyName, arguments);
+
+    amountGtThan0OrMinus1(target, propertyName, arguments);
+
+    return method?.apply(this, arguments);
+  };
+}
+
+export function LPSwapCollateralValidator(
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  target: any,
+  propertyName: string,
+  descriptor: TypedPropertyDescriptor<any>
+): any {
+  const method = descriptor.value;
+  // eslint-disable-next-line no-param-reassign
+  descriptor.value = function () {
+    const { LENDING_POOL, SWAP_COLLATERAL_ADAPTER } =
+      this.lendingPoolConfig || {};
+
+    if (
+      !utils.isAddress(LENDING_POOL) ||
+      !SWAP_COLLATERAL_ADAPTER ||
+      !utils.isAddress(SWAP_COLLATERAL_ADAPTER)
+    ) {
+      console.error(
+        `[LPSwapCollateralValidator] You need to pass valid addresses`
+      );
+      return [];
+    }
+
+    isEthAddressValidator(target, propertyName, arguments);
+
+    amountGtThan0Validator(target, propertyName, arguments);
+
+    amountGtThan0OrMinus1(target, propertyName, arguments);
+
+    return method?.apply(this, arguments);
+  };
+}
 
 export function LPValidator(
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
@@ -21,10 +118,10 @@ export function LPValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.lendingPool[this.market];
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const { LENDING_POOL } = this.lendingPoolConfig || {};
+
+    if (!utils.isAddress(LENDING_POOL)) {
+      console.error(`[LendingPoolValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -47,9 +144,11 @@ export function LTAMigratorValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] = enabledNetworksByService.ltaMigrator;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const LEND_TO_AAVE_MIGRATOR =
+      this.migratorConfig?.LEND_TO_AAVE_MIGRATOR || '';
+
+    if (!utils.isAddress(LEND_TO_AAVE_MIGRATOR)) {
+      console.error(`[MigratorValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -70,10 +169,14 @@ export function IncentivesValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.incentivesController;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const { INCENTIVES_CONTROLLER, INCENTIVES_CONTROLLER_REWARD_TOKEN } =
+      this.incentivesConfig || {};
+
+    if (
+      !utils.isAddress(INCENTIVES_CONTROLLER_REWARD_TOKEN) ||
+      !utils.isAddress(INCENTIVES_CONTROLLER)
+    ) {
+      console.error(`[IncentivesValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -94,10 +197,13 @@ export function LiquiditySwapValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.liquiditySwapAdapter;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const SWAP_COLLATERAL_ADAPTER =
+      this.swapCollateralConfig.SWAP_COLLATERAL_ADAPTER || '';
+
+    if (!utils.isAddress(SWAP_COLLATERAL_ADAPTER)) {
+      console.error(
+        `[LiquiditySwapValidator] You need to pass valid addresses`
+      );
       return [];
     }
 
@@ -120,10 +226,13 @@ export function RepayWithCollateralValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.repayWithCollateralAdapter;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const REPAY_WITH_COLLATERAL_ADAPTER =
+      this.repayWithCollateralConfig?.REPAY_WITH_COLLATERAL_ADAPTER || '';
+
+    if (!utils.isAddress(REPAY_WITH_COLLATERAL_ADAPTER)) {
+      console.error(
+        `[RepayWithCollateralValidator] You need to pass valid addresses`
+      );
       return [];
     }
 
@@ -146,10 +255,53 @@ export function StakingValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.staking[this.tokenStake];
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    // No need to check if addresses exist for network
+    // because this is checked at initialization and type checking of config
+
+    const { TOKEN_STAKING, STAKING_REWARD_TOKEN } = this.stakingConfig || {};
+
+    // Check if addresses are valid.
+    if (
+      !utils.isAddress(TOKEN_STAKING) ||
+      !utils.isAddress(STAKING_REWARD_TOKEN)
+    ) {
+      console.error(`[StakingValidator] You need to pass valid addresses`);
+      return [];
+    }
+
+    const isParamOptional = optionalValidator(target, propertyName, arguments);
+
+    isEthAddressValidator(target, propertyName, arguments, isParamOptional);
+
+    amountGtThan0Validator(target, propertyName, arguments, isParamOptional);
+
+    return method?.apply(this, arguments);
+  };
+}
+
+export function SignStakingValidator(
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  target: any,
+  propertyName: string,
+  descriptor: TypedPropertyDescriptor<any>
+): any {
+  const method = descriptor.value;
+  // eslint-disable-next-line no-param-reassign
+  descriptor.value = function () {
+    // No need to check if addresses exist for network
+    // because this is checked at initialization and type checking of config
+
+    const { TOKEN_STAKING, STAKING_REWARD_TOKEN, STAKING_HELPER } =
+      this.stakingConfig || {};
+
+    // Check if addresses are valid.
+    if (
+      !utils.isAddress(TOKEN_STAKING) ||
+      !utils.isAddress(STAKING_REWARD_TOKEN) ||
+      !STAKING_HELPER ||
+      !utils.isAddress(TOKEN_STAKING)
+    ) {
+      console.error(`[StakingValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -172,9 +324,10 @@ export function FaucetValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] = enabledNetworksByService.faucet;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const FAUCET = this.faucetConfig?.FAUCET;
+
+    if (!FAUCET || (FAUCET && !utils.isAddress(FAUCET))) {
+      console.error(`[FaucetValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -197,9 +350,10 @@ export function WETHValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] = enabledNetworksByService.wethGateway;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const WETH_GATEWAY = this.wethGatewayConfig?.WETH_GATEWAY || '';
+
+    if (!utils.isAddress(WETH_GATEWAY)) {
+      console.error(`[WethGatewayValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -222,10 +376,20 @@ export function GovValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.aaveGovernanceV2;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
+    const {
+      AAVE_GOVERNANCE_V2,
+      AAVE_GOVERNANCE_V2_HELPER,
+      AAVE_GOVERNANCE_V2_EXECUTOR_SHORT,
+      AAVE_GOVERNANCE_V2_EXECUTOR_LONG,
+    } = this.governanceConfig || {};
+
+    if (
+      !utils.isAddress(AAVE_GOVERNANCE_V2) ||
+      !utils.isAddress(AAVE_GOVERNANCE_V2_HELPER) ||
+      !utils.isAddress(AAVE_GOVERNANCE_V2_EXECUTOR_SHORT) ||
+      !utils.isAddress(AAVE_GOVERNANCE_V2_EXECUTOR_LONG)
+    ) {
+      console.error(`[GovernanceValidator] You need to pass valid addresses`);
       return [];
     }
 
@@ -246,13 +410,6 @@ export function GovDelegationValidator(
   const method = descriptor.value;
   // eslint-disable-next-line no-param-reassign
   descriptor.value = function () {
-    const currentNetwork = this.config.network;
-    const acceptedNetworks: Network[] =
-      enabledNetworksByService.aaveGovernanceV2;
-    if (acceptedNetworks.indexOf(currentNetwork) === -1) {
-      return [];
-    }
-
     isEthAddressValidator(target, propertyName, arguments);
     isEthAddressOrEnsValidator(target, propertyName, arguments);
     amountGtThan0Validator(target, propertyName, arguments);